MAINT: code refactoring Moved functions that fix tables into the process_markdowns.py script
